commit all
Karnaugh Map Minimizer is free (GPL) software for minimizing boolean functions using the graphic method of Karnaugh maps. I made it primarily for educational purposes so double check the output if you intend to use it for any "real" work. I hope it helps you with whatever you are doing.